Moved primary/queries out of the primary folder.
