+++ src/api/query/ant-multiple-result-query-manager.ts
@@ -1,7 +1,7 @@
 import { AntQueryManager } from './ant-query-manager';
 import { ApiMultipleResultQueryManager } from './api-multiple-result-query-manager';
 import { Entity } from '../../model/entity';
-import { QueryManager } from '../../persistence/primary/query/query-manager';
+import { QueryManager } from '../../persistence/query/query-manager';
 
 export class AntMultipleResultQueryManager<TEntity extends Entity> extends AntQueryManager<TEntity, TEntity[]>
   implements ApiMultipleResultQueryManager<TEntity> {
