+++ src/api/query/ant-query-manager.ts
@@ -2,7 +2,7 @@ import { AntJsSearchOptions } from '../../persistence/options/antjs-search-optio
 import { ApiQueryManager } from './api-query-manager';
 import { Entity } from '../../model/entity';
 import { PersistencySearchOptions } from '../../persistence/options/persistency-search-options';
-import { QueryManager } from '../../persistence/primary/query/query-manager';
+import { QueryManager } from '../../persistence/query/query-manager';
 
 export abstract class AntQueryManager<TEntity extends Entity, TResult extends TEntity | TEntity[]>
   implements ApiQueryManager<TEntity, TResult> {
