+++ src/persistence/query/ant-single-result-primary-query-manager.ts
@@ -1,9 +1,9 @@
 import * as _ from 'lodash';
 import { AntPrimaryQueryManager } from './ant-primary-query-manager';
-import { Entity } from '../../../model/entity';
-import { PersistencySearchOptions } from '../../options/persistency-search-options';
+import { Entity } from '../../model/entity';
+import { PersistencySearchOptions } from '../options/persistency-search-options';
 import { SingleResultPrimaryQueryManager } from './single-result-primary-query-manager';
-import { VOID_RESULT_STRING } from '../lua-constants';
+import { VOID_RESULT_STRING } from '../primary/lua-constants';
 
 export class AntSingleResultPrimaryQueryManager<TEntity extends Entity>
   extends AntPrimaryQueryManager<TEntity, number | string>
