+++ src/persistence/query/single-result-primary-query-manager.ts
@@ -1,5 +1,5 @@
 import { BasePrimaryQueryManager } from './primary-query-manager';
-import { Entity } from '../../../model/entity';
+import { Entity } from '../../model/entity';
 
 export interface SingleResultPrimaryQueryManager<TEntity extends Entity>
   extends BasePrimaryQueryManager<TEntity, TEntity> {}
