+++ src/test/primary/model-manager-test.ts
@@ -7,15 +7,15 @@ import { AntSchedulerModelManager } from '../../persistence/scheduler/ant-schedu
 import { CacheMode } from '../../persistence/options/cache-mode';
 import { Entity } from '../../model/entity';
 import { Model } from '../../model/model';
-import { MultipleResultQueryByFieldManager } from './query/multiple-result-query-by-field-manager';
+import { MultipleResultQueryByFieldManager } from '../query/multiple-result-query-by-field-manager';
 import { PrimaryEntityManager } from '../../persistence/primary/primary-entity-manager';
 import { PrimaryModelManager } from '../../persistence/primary/primary-model-manager';
-import { PrimaryQueryManager } from '../../persistence/primary/query/primary-query-manager';
+import { PrimaryQueryManager } from '../../persistence/query/primary-query-manager';
 import { RedisWrapper } from './redis-wrapper';
 import { SchedulerModelManager } from '../../persistence/scheduler/scheduler-model-manager';
 import { SecondaryEntityManagerMock } from '../../testapi/api/secondary/secondary-entity-manager-mock';
-import { SingleResultPrimaryQueryManager } from '../../persistence/primary/query/single-result-primary-query-manager';
-import { SingleResultQueryByFieldManager } from './query/single-result-query-by-field-manager';
+import { SingleResultPrimaryQueryManager } from '../../persistence/query/single-result-primary-query-manager';
+import { SingleResultQueryByFieldManager } from '../query/single-result-query-by-field-manager';
 import { Test } from '../../testapi/api/test';
 import { iterableFilter } from '../util/iterable-filter';
 import { iterableFind } from '../util/iterable-find';
