+++ src/test/query/single-result-query-manager-test.ts
@@ -1,17 +1,17 @@
-import { AntJsSearchOptions } from '../../../persistence/options/antjs-search-options';
-import { AntModel } from '../../../model/ant-model';
-import { AntPrimaryModelManager } from '../../../persistence/primary/ant-primary-model-manager';
-import { AntSchedulerModelManager } from '../../../persistence/scheduler/ant-scheduler-model-manager';
-import { Entity } from '../../../model/entity';
-import { Model } from '../../../model/model';
-import { PrimaryModelManager } from '../../../persistence/primary/primary-model-manager';
-import { RedisWrapper } from '../redis-wrapper';
-import { SchedulerModelManager } from '../../../persistence/scheduler/scheduler-model-manager';
-import { SecondaryEntityManager } from '../../../persistence/secondary/secondary-entity-manager';
-import { SecondaryEntityManagerMock } from '../../../testapi/api/secondary/secondary-entity-manager-mock';
+import { AntJsSearchOptions } from '../../persistence/options/antjs-search-options';
+import { AntModel } from '../../model/ant-model';
+import { AntPrimaryModelManager } from '../../persistence/primary/ant-primary-model-manager';
+import { AntSchedulerModelManager } from '../../persistence/scheduler/ant-scheduler-model-manager';
+import { Entity } from '../../model/entity';
+import { Model } from '../../model/model';
+import { PrimaryModelManager } from '../../persistence/primary/primary-model-manager';
+import { RedisWrapper } from '../primary/redis-wrapper';
+import { SchedulerModelManager } from '../../persistence/scheduler/scheduler-model-manager';
+import { SecondaryEntityManager } from '../../persistence/secondary/secondary-entity-manager';
+import { SecondaryEntityManagerMock } from '../../testapi/api/secondary/secondary-entity-manager-mock';
 import { SingleResultQueryByFieldManager } from './single-result-query-by-field-manager';
-import { Test } from '../../../testapi/api/test';
-import { iterableFind } from '../../util/iterable-find';
+import { Test } from '../../testapi/api/test';
+import { iterableFind } from '../util/iterable-find';
 
 const MAX_SAFE_TIMEOUT = Math.pow(2, 31) - 1;
