+++ src/testapi/api/generator/model-manager-generator.ts
@@ -1,20 +1,20 @@
 import * as _ from 'lodash';
 import { AntJsSearchOptions } from '../../../persistence/options/antjs-search-options';
-import { AntMultipleResultPrimaryQueryManager } from '../../../persistence/primary/query/ant-multiple-result-primary-query-manager';
-import { AntSingleResultPrimaryQueryManager } from '../../../persistence/primary/query/ant-single-result-primary-query-manager';
+import { AntMultipleResultPrimaryQueryManager } from '../../../persistence/query/ant-multiple-result-primary-query-manager';
+import { AntSingleResultPrimaryQueryManager } from '../../../persistence/query/ant-single-result-primary-query-manager';
 import { ApiModelManagerGeneratorOptions } from './api-model-manager-generator-options';
 import { ApiModelManagerGeneratorRedisOptions } from './api-model-manager-generator-redis-options';
 import { ApiModelManagerGeneratorSecodaryManagerOptions } from './api-model-manager-generator-secodary-manager-options';
 import { ApiQueriesManagerGeneratorOptions } from './api-queries-manager-generator-options';
 import { Entity } from '../../../model/entity';
 import { Model } from '../../../model/model';
-import { MultipleResultPrimaryQueryManager } from '../../../persistence/primary/query/multiple-result-primary-query-manager';
+import { MultipleResultPrimaryQueryManager } from '../../../persistence/query/multiple-result-primary-query-manager';
 import { PrimaryModelManager } from '../../../persistence/primary/primary-model-manager';
-import { PrimaryQueryManager } from '../../../persistence/primary/query/primary-query-manager';
+import { PrimaryQueryManager } from '../../../persistence/query/primary-query-manager';
 import { RedisMiddleware } from '../../../persistence/primary/redis-middleware';
 import { SchedulerModelManager } from '../../../persistence/scheduler/scheduler-model-manager';
 import { SecondaryEntityManager } from '../../../persistence/secondary/secondary-entity-manager';
-import { SingleResultPrimaryQueryManager } from '../../../persistence/primary/query/single-result-primary-query-manager';
+import { SingleResultPrimaryQueryManager } from '../../../persistence/query/single-result-primary-query-manager';
 
 export abstract class ModelManagerGenerator<
   TOptions extends ApiModelManagerGeneratorOptions<
